Introduce CompoundedSet for first sets and follow sets
A compounded set is implemented internally as a local set, and a set of other
sets whose members are also considered to be within the compounded set. This is
useful when said other sets are mutable, as modifying them will modify all
containing compound sets. Thus compounding a set into a compounded set
establishes a /permanent/, /algebraic/ relationship between the sets, where the
compounded set is always a superset of the new set.

Signed-off-by: Casey Dahlin <casey.dahlin@gmail.com>