Fix the notorious "Zero-length bug"
We are actually susceptible to a much worse form of this, since our state sets
are keyed on text position, not token position (as required by our tokenization
model). The gist of the issue is "what if a state is completed, and then a
state that could have been completed is inserted?"

We now store completing states so they can be retried if their target set
expands.

Signed-off-by: Casey Dahlin <casey.dahlin@gmail.com>