Don't bind tokenizers multiple times
Previously, out of frustration, we would turn a tokenizer to a declaration
every time we called it and then use memberInvoke. This stepped around type
safety as best as we could, but was very expensive as there was lots of
metamodel execution.

This method stems from a less naive understanding of the metamodel, and binds
exactly the functions we want at compile time.

Signed-off-by: Casey Dahlin <casey.dahlin@gmail.com>