Implement the new tokenization API
We're doing away with tokenizer functions. Instead, token objects are now
expected to know how to produce the next available token. Token is now an
interface and the user can implement his tokenizer by implementing token.

Grammar no longer has a data type and no longer takes a list. Instead we pass
it an SOSToken (start-of-stream token) which represents the beginning position
in the string and has been overridden to have the appropriate `next` method.

Signed-off-by: Casey Dahlin <casey.dahlin@gmail.com>