Extract shortcut function to reduce generated code
Instead of TokenizerToken<X>, we can now write tokenizerToken<X>() in
the map. Since this is not a callable reference, but a regular function
call, the compiler no longer has to generate an anonymous
`AbstractCallable` subclass – all it has to do is reify the type
argument (for a concrete class, `Given.$TypeDescriptor$`).

Signed-off-by: Lucas Werkmeister <mail@lucaswerkmeister.de>