Wrong equals() method in CeylonToken?
Looks to me as the `equals()` method of `CeylonToken` is wrong. You wouldn't want to equal two tokens that are named the same and which are in the same line but in a different column position. Maybe you intented to write `text && line_start && col_start` instead of `text && line_start && line_end`. But I might be wrong...