Rework StateQueue
We're now using a more conventional implementation for the state sets,
consisting of a single linear array. We have to use a priority queue for the
next state set, since we could potentially affect more than one successor
queues due to the fact that we key on text position rather than token stream
position. We also have a novel method of dealing with the zero-length bug.

Signed-off-by: Casey Dahlin <casey.dahlin@gmail.com>